Split input into files
The file was split because the number of input modes increased.
